Don't let entries in the persistent symbol table become too large leading to assertion/crash, instead give out a warning.

+++ language/duchain/persistentsymboltable.cpp
@@ -119,7 +119,7 @@ class PersistentContextTableItem {
     return id.index;
   }
   
-  unsigned short int itemSize() const {
+  size_t itemSize() const {
     return dynamicSize();
   }
   
@@ -200,6 +200,11 @@ void PersistentSymbolTable::addDeclaration(const IndexedQualifiedIdentifier& id,
       item.declarationsList().append(oldItem->declarations()[a]);
     }
     
+    if(request.itemSize() > (1<<15)) {
+      kWarning() << "too many declarations for id" << id.identifier().toString() << item.declarationsSize() << "ignoring new one";
+      return;
+    }
+    
     d->m_declarations.deleteItem(index);
   }
 
@@ -278,6 +283,10 @@ void PersistentSymbolTable::addContext(const IndexedQualifiedIdentifier& id, con
       item.contextsList().append(oldItem->contexts()[a]);
     }
     
+    if(request.itemSize() > (1<<15)) {
+      kWarning() << "too many contexts for id" << id.identifier().toString() << item.contextsSize() << "ignoring new one";
+      return;
+    }
     d->m_contexts.deleteItem(index);
   }
